Started to reorganize the customer page to include the reservation
