+++ client/src/components/CustomerReview.js
@@ -1,38 +1,58 @@
-import Auth from '../utils/auth';
-import { ADD_REVIEW } from '../utils/mutations';
-import { useState } from "react";
-import {useMutation} from '@apollo/client';
+import { useQuery, useMutation } from "@apollo/client";
+import { GET_USER} from "../utils/queries";
+import {DELETE_REVIEW} from "../utils/mutations";
+import Auth from "../utils/auth";
+import { OverflowMapped,  CustomerMappedReviews} from "../styled/Customer.styled";
 
-const CustomerReviewSection = () => {
-        //For ADD_REVIEW 
-        const [reviewText, setReviewText] = useState('');
-        const [addReview] = useMutation(ADD_REVIEW);
+const CustomerReviews = () => {
+        
+    //For the GET_USER:
+    const {loading, data} = useQuery(GET_USER);
+    const userData = data?.user || {};
 
-        const handleSavedReview = async() => {
-        const token = Auth.loggedIn() ? Auth.getToken() : null;    
-        if (!token) {
-          return false;
-        }
-        try{
-            const {data} = await addReview({
-             variables: {
-                reviewText,
-                reviewAuthor: Auth.getProfile().data.username,}
-            });
-            console.log(data);
+    const {reviews} = userData;
+   
+    if(loading){
+        <h2>Loading...</h2>
+    }
+     //For DELETE_REVIEW
+     const [deleteReview] = useMutation(DELETE_REVIEW);
 
-        } catch(error){
-            console.error(error);
+    const deleteAReview = async(reviewId) => {
+    const token = Auth.loggedIn() ? Auth.getToken() : null;
+
+
+     if (!token) {
+     return false;
         }
-        }
+
+     try{
+
+     const {data} = await deleteReview({
+     variables: {reviewId}
+
+     });
+     console.log(data);
+     window.location.reload();
+    } catch(error){
+          console.error(error);
+    } }
 
     return (
-        <form onSubmit={handleSavedReview}>
-        <label>Leave a review:</label>
-        <textarea name='reviewText' value={reviewText} onChange={(e)=> setReviewText(e.target.value)} rows="4" cols="50"/>
-        <button className="button">Submit</button>
-       </form>
-      );
-}
- 
-export default CustomerReviewSection;
+        <>
+         {reviews ? (
+            <OverflowMapped>
+            {reviews.map(review => (
+                <CustomerMappedReviews key={review.reviewId}>
+                <p>{review.reviewText}</p>
+                <button className="button" onClick={() => deleteAReview(review._id)}>Delete</button>
+                </CustomerMappedReviews>
+            ))}
+            </OverflowMapped>
+        ) : (<h6>You have no reviews</h6>)}
+        
+        </>
+    )
+};
+
+export default CustomerReviews;
