+++ client/src/pages/Customer.js
@@ -1,49 +1,20 @@
-import { useQuery, useMutation } from "@apollo/client";
+import { useQuery} from "@apollo/client";
 import { GET_USER} from "../utils/queries";
-import {DELETE_REVIEW} from "../utils/mutations";
-import { OverflowMapped, StyledCustomer, StyledCustomerRight, CustomerMappedReviews, CustomerReview } from "../styled/Customer.styled";
+import { useState } from "react";
+import {StyledCustomer, StyledCustomerRight, CustomerReview } from "../styled/Customer.styled";
 import img4 from '../images/food/img4.jpg';
 import Auth from "../utils/auth";
-import CustomerReviewSection from "../components/CustomerReview";
+import CustomerReviewSection from "../components/CustomerReviewForm";
 //import EditReview from "../components/EditReview";
-//import { Link } from "react-router-dom";
+import { Link } from "react-router-dom";
 
 const CustomerPage = () => {
-    
-    //For the GET_USER:
     const {loading, data} = useQuery(GET_USER);
     const userData = data?.user || {};
 
-    const {reviews} = userData;
-   
-    if(loading){
-        <h2>Loading...</h2>
-    }
-
-     //For DELETE_REVIEW
-     const [deleteReview] = useMutation(DELETE_REVIEW);
-
-    const deleteAReview = async(reviewId) => {
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
-
-
-     if (!token) {
-     return false;
-        }
-
-     try{
-
-     const {data} = await deleteReview({
-     variables: {reviewId}
-
-     });
-     console.log(data);
-     window.location.reload();
-    } catch(error){
-          console.error(error);
-    } }
+    const [currentSection, setCurrentSection] = useState('Reviews');
    // If we ever put in the edit again  <EditReview reviewId={review._id}/>
-   //The link for seeing all your reservations <Link to='/customerReservations'>Your Reservations</Link>
+   //The link for seeing all your reservations 
     return ( 
         <>
     {Auth.loggedIn() ? (
@@ -52,22 +23,14 @@ const CustomerPage = () => {
        <StyledCustomerRight>
         <CustomerReview>
         <h2>{userData.username}</h2>
+        <div>
+        <Link to='/customerReviews'><button className="button">Reviews</button></Link>
+        <Link to='/customerReservations'><button className="button">Reservations</button></Link>
+        </div>
         <CustomerReviewSection />
         </CustomerReview>
-        {reviews ? (
-            <OverflowMapped>
-            {reviews.map(review => (
-                <CustomerMappedReviews key={review.reviewId}>
-                <p>{review.reviewText}</p>
-                <button className="button" onClick={() => deleteAReview(review._id)}>Delete</button>
-                </CustomerMappedReviews>
-            ))}
-            </OverflowMapped>
-           
-           
-        ) : (<h6>You have no reviews</h6>)}
-       </StyledCustomerRight>
-      
+       
+       </StyledCustomerRight>   
     </StyledCustomer>
     
     ) : (
